New Development: no             JIRA Issue: no       Ready to Release: yes      Interface Changes: none What Interface Changed: none          Test Programs: none   Put in Release Notes: no
           Description: giving up... will try again when I've worked out the RedHat issues...



git-svn-id: https://svn.cv.nrao.edu/svn/casa/active@18157 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ code/display/implement/region/Point.cc
@@ -59,7 +59,7 @@ namespace casa {
 	    double wblc_x, wblc_y;
 	    linear_to_world( wc_, blc_x, blc_y, wblc_x, wblc_y );
 
-	    double pblc_x, pblc_y;
+	    int pblc_x, pblc_y;
 	    linear_to_pixel( wc_, blc_x, blc_y, pblc_x, pblc_y );
 
 	    pixel_pts.resize(1);
